fix: fetch all of known-issues docs
